Change sgit -o (sed options) to be an array
This allows sgit -o to be used more than once.

This commit slightly changes the usage of -o: now one must specify - or
(for long options) --. This is an improvement and a bug fix. Originally
an oversight, it only allowed for the use of -o once (or rather each
time it was used it was overwritten). That meant that one had to quote
the options list but worse than that only one '-' was used which for
more than one option is obviously an error.

Now one can specify more than one option either by one -o (quoting the
option arg) or by using -o more than once. Not all options of sed have
been tested and this will never be done by me.

New version 0.0.9-1 24-04-2023.

Significant improvements to README.md.

Updated top comments in sgit as well.