Improve fix of 1990/tbr
In particular the fix to make it more like the original. There was no
need to have the function prototypes at the top of the file.

I also added a couple -Wno- options to the CSILENCE variable in the
Makefile.

I also added to the Makefile so that make alt will build the alt version
(only functional difference is that there is gets() instead of fgets()).