Add sgit.1 man page, minor fixes in sgit
Something that is lacking from the man page is that there are no
examples. This is currently being justified by the fact that the
README.md has examples but since that might not always be available this
doesn't hold very well. This will likely be fixed another time when I
have more energy, time and patience for it.

As it is not clear to me how any documentation should be 'installed' I
have not updated the Makefile to address this. That can be done later -
or not for that matter.

The usage string has been improved a bit. There was one thing that could
be more clear and there was something missing too.

Fixed exit codes when command line error.

README.md now refers to the man page and the usage string was updated to
match the usage string change.

New version 0.0.11-1 28-04-2023.