Add ioccc_status.sh to quickly update status.json
This was done quickly to give me something to do but if it's useful in
any form then so much the better but it seems it's more ideal to update
the status.json file automatically rather than having to do it manually.

The script does not have to update every field in the status.json file.
It doesn't even have to update any for that matter. The usage is:

    usage: ioccc_status [-h] [-V] [-v level] [-s status] [-d] [-n] [-I status_ver] status.json

	-h			    print help and exit
	-V			    print version and exit
	-v level		    set verbosity level
	-s statue		    set contest status to open or closed
	-d			    update status_date
	-n			    update latest_news date
	-I status_ver		    update IOCCC_status_version

					NOTE: status_ver must be one of "open" or "closed"

	status.json		    the file to update

    status version: 0.0.1-0 2023-10-02

Again if this is not useful that's perfectly okay: it gave me something
to do and it seems useful but there certainly are other ways that it
could be done, better, worse or just different.

Obviously this script does not first verify that it is valid json or the
correct format. If the format is incorrect it will not be updated.

Added to tmp/src/ioccc_status and updated the tmp/src/Makefile.

Added to tmp/src/ioccc_status/.gitignore ioccc_status: the file in the
repo is ioccc_status.sh.