Add alt code to 1993/rince
This is based on the suggestion of the judges that some people might
want to slow the game down a bit.

It is easy to configure at compile time like:

    make CDEFINE=-DZ=18 clobber alt

More details are in the README.md file.

Updated .gitignore and thanks file (not just for this but also
making something else clearer in another entry).