Force regexp in IOCCC_status_version
As it seems like the version should be a specific format both for
consistency and to keep track of it I added a regexp check in the
ioccc_status.sh script I wrote the other day that does this.

The version must follow the regexp that was devised from the current
IOCCC_status_version of '0.2 2023-05-22' as:

    [0-9]+\.[0-9]+ [0-9]{4}-[0-9]{2}-[0-9]{2}

The usage string has had this added and a new check was added to verify
that the '-i version' matches this format.

This update also will prevent sed from reporting an error if the
separator character (which I originally chose as '/' and then changed it
to '|' but which now could be either or ',' or something else) is in the
version as the version won't have that character.

The format of the version can certainly be changed but this seems like a
worthy idea, to check that there is a required format.

With this the script should be finished though of course something might
be found later on.