Silence compilation of 1985/sicherman
A warning that might not have been previously triggered, odd and
unlikely as it might seem, was triggered at least in macOS Sonoma and it
has now been disabled: -Wno-implicit-int.

The entry used to require in modern days -traditional-cpp but like the
others I fixed this and that did introduce new warnings as code had to
be commented out in some cases like this entry (it did not matter in
this entry if it was commented out but it had to be due to compiler
errors otherwise). Even so this warning should have been triggered
already due to main() and subr() not specifying a return type.