Fix links in thanks-for-fixes.md wrt winners.html
Since linking to html files on GitHub does not render the html I have
changed the winners.html to be https://www.ioccc.org/winners.html
instead.

In the process I have discovered that several README.md files still
include '/winners.html' and these include some that have both Yusuke and
me. Some of the names are supposed to stay there as they're mentioned by
the author but some are not. These will be fixed in time: I have other
modifications including to some README.md files and I don't want to
accidentally commit the wrong files as some have changes that are not
necessarily finished. Since it's the same kind of change it feels like a
single pull request would be fitting (and easier) but since all the
README.md files will have to be checked again (some once, some twice and
that should be it) I will make these fixes (for those that need to be
fixed .. for instance I can think of at least three where I was
mentioned by another author so those should NOT be removed but the link
should be changed!) then.

I have updated the todo.md file as well to note the above as before I
can say everything is done with the current issues related to
consistency and typos the file must be checked and everything ticked off
(or deleted) as the tasks are done.