Typo fix and warning silence in 1988/applin
The rule zsmall.c comment suggested that applin rule (which is done via
the ${PROG} variable which is the value of ${ENTRY} which is how every
other entry Makefile is done) was below but actually it is above.

If ${CC} is clang then add to CSILENCE -Wno-gnu-line-marker. This
unfortunately means one has to either have by default CC as clang or
they have to specify e.g.:

    make clobber CC=clang everything

as CC tends to be just 'cc'.

The warning cannot be silenced with gcc and although it allows one to
use of -Wno-gnu-line-marker it has no effect other than to say it's
unrecognised because gcc (at least in some versions?) does not allow
this warning to be silenced, for whatever (invalid :-) ) reason(s). Thus
unless one explicitly says that CC=clang they will get the warning.

Still at least it's now silenced if one does indeed do CC=clang in the
compile line.