Update bugs and thanks for 1992/lush
I also removed a -Wno- option from the Makefile as the entry relies on
this warning. This is not strictly necessary as the script that I added,
runme.sh, directly compiles it, but it felt like it shouldn't have it
anyway.

In the bugs.md file I gave some hints as to why this does not work with
clang.

In the thanks file I broke up a paragraph into more than one paragraph.