Add important note to 1989/fubar/README.md
In the judges' remarks they refer to ouroboros.c as being an executable
but it's not in the repo. This is because it's created by the program
itself. It happens to be in .gitignore but this note should make it less
confusing and prevent people from thinking that a file is actually
missing.