Swap 2006/sloane alt/main code in build/try
As the original entry is too fast for modern systems I have (after
discussion with Landon) made the build and try sections refer to
'sloane.alt' instead of the original entry. The 'alternate section' now
refers to 'sloane' instead. This is a rare case where the reversal of
alt (in this case provided by me) and the original is valuable.

There are other entries where I will end up doing this as well for the
same reason of the entry going too fast in modern systems and these will
be updated in time, now that a method to do it has been established.