Add 2>/dev/null to try in 1986/hague
Until such a time that the alt code can become the main code, to be
safer and more importantly to prevent obnoxious warnings that are shown
at runtime, the 2>/dev/null will prevent the runtime warning from
showing up; this is particularly 'necessary' for macOS as it is the one
that shows the warning and this entry is more complicated to set up
fgets() though I do have some ideas on how it might be done but 'might'
is a somewhat big if: it's probably in this case going to be tampering
too much with the entry when the benefit is not that worth it since we
can redirect the warning to /dev/null. This is TBD later.