Swap 2013/birken alt/main code in build/try
As the original entry is too fast for modern systems I have (after
discussion with Landon) made the build and try sections refer to
'birken.alt' instead of the original entry. The 'alternate code' section
was changed to 'original code' and it refers to 'birken' instead.

This is a rare case where the reversal of alt (in this case provided by
me) and the original is valuable.

There are still some other entries, I believe, where this will be useful
to do as well and for the same reason as this entry: it goes too fast in
modern systems.

The demo.sh script that I wrote now defaults to birken.alt.

Minor fixes to thanks-for-fixes.md wrt this entry.