Fixes to 2013/cable1/README.md
Although I already checked formatting the tweet of Landon's that was
referred to was not on the Internet Wayback Machine so instead I took a
screenshot of the tweet and embedded it into the README.md file so that
when generated as a html file it will show the tweet instead (dark mode
because I don't like light backgrounds).

BTW: the author said: 'prejudge the past' but how can one prejudge the
past? I suggest a new word: postjudge :-)

(No this does not technically ignore the definition as it refers to
PRE-maturely AND without adequate information :-) but even if it did
that's far less fun! Of course it could be judging before seeing it but
this is also much less fun :-) )