Further compilation fix to 2013/cable3
Whereas in earlier versions of macOS my previous changes let it compile
in at least the most recent version, macOS 14 (Sonoma), it did not due
to the same problem of localtime(). The prototype has been removed
entirely to solve the problem.

The runme script no longer assumes that the program was compiled,
running first:

    make clobber all || exit 1

I partly formatted the README.md file. I will have to finish this file
and continue to the rest of 2013 another time.

One thing of importance (I feel) that I did add is that I put at the top
that to quit the emulator one should type QUITEMU as this was buried
deep in the remarks.