Add alt code to 2013/cable3 + other fixes
The alt code _should_ (keyword :-) ) run in Windows. This code is based
on the author's remarks. It is done by (at the compiler line) undefining
KB and defining it in the source code (cable3.alt.c) as the author
suggested.

The cable3.c now #includes unistd.h: this is because the author
suggested in some systems it might be necessary though the example the
author gave no longer exists for current versions (the last 32-bit macOS
to support 32-bit binaries was version 0.14 Mojave and even that was
64-bit and Mojave was released in 2018 with a new version released each
year since then). This does trigger a warning in macOS but we now
silence that warning as well.

The runme script now passes shellcheck (make '[ .. ]' '[[ .. ]]').

Format fixes in README.md including a few links where this seemed like a
good idea (for instance websites).

Updated .gitignore: added cable3.alt.

As as of yesterday there is no need to #undef localtime any more (if
there ever was: the fix I made to let it compile did not in the end need
it but it was the first thing I had done so I kept it in) this has been
removed as well.

The screenshots provided by the author have been linked into the
README.md file so that they should display in the html version.

At least one file was changed in name (in the README.md file) to what I
believe is the correct name (it was a non-existent file).

Updated bugs.md to refer to a file that is not found that the author
referred to.

Add bios.asm that author referred to: found at the GitHub repo for the
entry which the README.md now links to (was another website that no
longer exists).