Format fixes and additions to 2013/endoh4
Format and typo fix 2013/endoh4/README.md.

Based on the author's suggestions I added a script run.sh which
temporarily disables the cursor, using trap to set it back (it also
explicitly sets it back in case the signals specified are not sent to
the program). If no arg is specified it will feed the source code of the
program into the program; otherwise it'll use the arg specified. It does
not check if the file exists and is readable or anything else as the
shell/program will do that.