Format fixes to 2013/robison/README.md
To be more like the author's remarks I made the judges' remarks EOLs end
with spaces so that it hopefully is not merged. This is because it's in
verse. This matches what the author did as well.

A few links were added where it seemed useful or interesting.

I believe this file is good and need not be checked for typos later but
it's also hard to be sure because the way it is written (verses) some
things are not words and so I probably will go through it one more time.

Nevertheless with this commit the README.md files of the 2013 entries
(well, the entries of 2013, not 2013 entries! :-) ) have been format
checked and fixed.