Typo check entries for 1984
This commit includes the two entries not already done (other changes
besides a final typo and formatting pass) in 1984, anonymous and laman.

This commit means that the entries of 1984 have each had their README.md
file formatted and typo fixed one more time and this should be good for
publication.