Format/typo check 1985/sicherman/README.md
Unless (as seems possible) a format/style change is needed for alt code
sections, this should be the final pass over it.

The thanks-for-fixes.md file was fixed to not refer to the alt code for
the original but rather refer to sicherman.orig.c. The alt file was not
removed, however, because it is not yet clear how to deal with the
situation where alt code is the same as the original code and in
particular wrt Makefile rules. This is another reason that the README.md
file might need to be revisited again.