Update thanks file wrt 2014/skeggs + add -ldl
When I got this entry to work apparently I never added a thanks for it
so this is now done.

Also to make the compilation more portable I added -ldl to the LDFLAGS.
This is something Yusuke Endoh did and although not necessary in the
systems I tried it since it uses dlsym() it's better to have it.