Remove useless use of cat in 1987/biggar/README.md
This was in the 'Try' section but as a ailurophile (cat lover) I have to
say that if one removes a useless use of cat then literally speaking if
there were any cats left then they would be useful. I utterly object,
therefore, that there is no longer a cat here because cats most
certainly are useful. For instance if Earth was flat cats would have
pushed everything off it by now which is proof to those not in the Flat
Earth Society that the Earth is not flat. It's not proof to the Flat
Earthers but that's because they think it's flat regardless when in
reality the only related thing that might be flat is their brain, kind
of like a battery that's gone flat. And that's flat :-) But I digressed;
the point is that there was no need for the cat command here even if
there's always a use for cats. :-)