Improve how 1986/wall alt code is built
Add explicit rule to compile alt code and update CFLAGS to include (if
CC is gcc) -traditional-cpp. Clang does not support it. If CC is cc it
might support it depending on the system (macOS gcc is actually clang so
it won't but linux it will). This allows just running instead of:

    make CFLAGS+="-traditional-cpp" alt

to do:

    make CC=gcc alt

but still allowing for the former which might be necessary if CC is not
defined to be gcc.

The README.md points out that in macOS gcc is clang.

Improve the command (in the thanks file) of how to see the difference
between the -traditional-cpp version and the fixed version so that it
doesn't need git but rather just diff (as git diff is unnecessary here).
Note that there is discussion about new make rules that will make this
type of thing not needed so the thanks file would be updated at this
point for that to just refer to the FAQ (or not say anything about it at
all).