Format/typo check 1988/isaak/README.md
Added a section with historical notes about the entry (that is I added
the section heading and put the respective text, some of which I
previously added, under it).

This file should be in good order except perhaps a pending issue of
formatting the alternate code/to build/try sections.