Format/typo check 1988/applin
The README.md and the Makefile were both updated. The Makefile because
of typos/unclear things. The README.md had formatting fixes and also
added the historical note about 1988 from the Makefile as it's far less
likely people will see that note versus that which is in the README.md.