Improve 1988/westley/demo.sh
Check for executable 'westley' before compiling and only _after_
compilation (if necessary) is the source shown. This way if it's
already done (as likely the case) the result of the program will be
shown closer to the source code which is entirely the point. But even if
it's not already compiled since it's compiled first it'll have the
result closer to the source (if it fails though it'll not show the
source at all).