Remove me from 1990/baruch/README.md thanks
This was the last remaining missed case of me being in another README.md
which happened before the (I am a broken record, perhaps, but it's true)
vastly superior thanks-for-fixes.md file was thought of.

There are other entries not mine that refer to me but these come from
the authors themselves, not fixes, so they should stay the same.

I still have to check if there are any README.md files that have
references to Yusuke Endoh (for fixes) as well but that will be done in
another commit another time.

I updated the alternate code section and subsections here (the
formatting and text) as well as added at the top that there is an alt
version.