Fix 1989/tromp alt code
Originally I put in the author's code without fixes in place. This is
one of the reasons I did not see the improvements that he suggested.
Several fixes had to be made and now it works. It should be noted I have
not tested more than one user updating the high score file but the file
is now created (I had to make a similar fix to the original entry as
well).

Format/typo check README.md file. Not sure I'm done with this one though
as presentation might be improved.