Format fixes/typos in 2020/ferguson[12]
In part due to the adding \ at end of lines with spaces but there was a
typo on my part in a link.

Plus a minor update to the obfuscation files for ferguson2.

The files have not been checked thoroughly. I just happened to notice
these issues and wanted to correct them. I will dot hat later on,
assuming that I can get myself to go through the textbook-like files.