Format/typo check 1990/cmills
I included a link in the judges' remarks written by Dennis Ritchie about
C history in order to explain something the author noted, that I felt
some might find interesting. It was stated:

    ... as well as the irritating compiler I was using which assumes
    x=*p++ means x*=p++ !

but the explanation is quite simple. In early C and also B the compound
assignments had the = first so that:

    x += 5

used to be written as:

    x =+ 5

This is also shown in Lions' Commentary on the Sixth Edition UNIX
Operating System but not linked, only referenced. Both the document and
the book are quite fascinating.