Make my fix to 1993/vanb more like original
Now it's just a single line difference: main() itself that calls O5()
(which used to be main() via -DO5=main).

This was accomplished by using the -include stdio.h in the Makefile and
then removing (at the top of the file) the int O3. Instead main() (which
used to be O5() as noted and which had three args) has it as a char **.
Now since O3 was at file scope before it was 0 so rather than casting O3
in main() to an int the call to O5() just uses 0 in that place. The
moving the arg to main() is done simply because O5() (which again was
main()) also has 3 args.

Some versions of clang object to four args to main() but they allow 0,
2 or 3 so that's fine.