Test new concept of (mis)features/bugs in README
To make sure that this looks okay before doing it for the rest of the
entries I am doing this test commit to make sure the link works okay and
the format looks okay. If any problems are found (which I expect) they
will be corrected and once things look okay the rest of the files will
be done as well.