Improve links in thanks-for-fixes.md
Rather than have the specific entries noted at the top (for Yusuke and
me) link to the README.md file of the entries, it seems better to me to
link to the entry in the file as that explains what was done which is
what is being honoured/appreciated.

I also added another section or two (both short). One has a joke in it.

Some of the entries had improvements in text of what had to be done to
fix them to work / what was wrong.