Fix 1995/vanschnitz/README.md about LEVEL
For some reason when I tried it I had a problem with LEVEL being a
certain amount but it must have been something else in the system at the
time. Naturally modern systems are much more capable so it can accept a
much higher level before it gets slower.

Even the judges' statement of 15 causing a slow/bad time is not a
problem so I reworded it to remind readers that this was back in 1995
and is no longer the case.