Add scripts to 2012/kang
Add try.sh which runs the judges' try commands, showing what the
commands are and then the output, to simplify trying the entry out.

The scripts en.alt.sh and de.alt.sh are like en.sh and de.sh except it
defaults to the kang.alt version, which as noted can help one appreciate
the entry even more. Although not documented one can specify which
version for the alt scripts and the other scripts.