complete manifest by adding missing entries
The manifest now matches the files found in the winning
directories.

The tool, gen_file_list.found.sh, does a make clobber and
generates the list of files found in winning directories.

The tool, check_file_list.sh, may be used to verify that the manifest
matches the list of files found in winning directories.