Temporary (? :-) ) fixes in tmp/ scripts
Typos and make the renaming of check_awards.sh -> awards_check.sh
consistent in the comments (as in change the name of the script in the
comments to the new name).

None of these are super important as they're temporary files but they
are still incorrect so have to be fixed in my mind.