Make my fix to 2001/coupard more like original
In linux clang does not support implicit int and it's an error. I ended
up adding 'int' but to make it more like the original I added to the
CSILENCE -Wno-implicit-int instead. Thus the code is the original and
only the Makefile is different.

The thanks file was updated for this but also other entries (nothing
changed in the entries themselves, just changes in the description in
the thanks file).