Add files from 2020/endoh2 password-protected zip
The password was no longer known as the link that had a list of
passwords as provided by Yusuke in 2020 was no longer there, being
instead from 2021. I didn't have to use the Internet Wayback Machine as
I had already extracted them in the 2020 preview back then. The
spoiler/spoiler.md has been format fixed.

The spoiler.zip file has been deleted. The README.md file has been fixed
to refer to the directory, not the zip file, and a missing '`' was added
to the file that prevented proper rendering of the html.