Add FAQ about mismatch of source/author remarks
This felt important because numerous times I have had to put in a
README.md file and/or the thanks file that state that the authors' or
author's remarks no longer apply and what should be done about it. These
should be located and either be removed or be synchronised (or even
refer to the FAQ?).