Update bug status of 1989/westley
It appears it might not possible to fix the second and third version
that is generated by itself to compile with clang. It _might_ be but it
takes heavy modification. What's more, when I almost had it, it then
caused compilation errors in a different compiler (gcc). I now have a
good idea how to make the code create different functions (see thanks
file for more details on what had to be done to fix as much as possible)
and names and everything else which I might use in the future but for
now it is as good as it can get.

There is a known problem with version 2, namely that SOMETIMES it
segfaults after printing the output. It is hard to know if this should
be fixed but given the complications of this entry (it is truly a
masterpiece as was typical of Brian Westley) fixing it could also break
something else so I have changed it to INABIAF status.