Fix 1984/laman if no arg specified
As the definition of INABIAF changed, again, and since it took no more
than ten seconds to fix, I have fixed this entry.

As the thanks file was updated for this some other things that had been
left out were also fixed at the same time as these were already in the
file but not committed. There was also a typo fix in there.