Retitle a bug status about some compilers
Rather than have:

    STATUS: doesn't work with some compilers - please provide alternative code

It might be better to have:

    STATUS: doesn't work with some compilers - please provide alternative code or fix for all compilers

which it now is.

Only one entry had it in so that entry and bugs.md were both updated.
Another entry in bugs.md was modified before and the README.md but the
status was not changed. That will be in the next commit which will be
this retitled one.