Change bug status of 1996/august
Now it's considered a bug that might be worth fixing, either as alt code
or for changing it to work with more compilers.

I also renamed the status recently renamed just a tad: don't say fix for
all compilers but more compilers as it's highly unlikely one can test
every single compiler. This is the status this entry is classified
under.

Slight update to thanks file about this entry - make it a simpler note
about how some compilers (after fixing it to work at all which I did
quite some time back) will still compile it into an infinite loop but
it's a known issue. Clang in linux, incidentally, does not have the
problem but clang (which is even what /usr/bin/gcc is) in macOS does.

Other fixes in the bugs.md file were made as well particularly updates
to the explanation of some of the statuses at the top of the file.