Swap alt/original in 1995/cdua
It is recommended that with 1995/cdua one should start out with the
alternate version as the original goes too fast in modern systems to see
it do its task. The alternate code section has been changed to original
code.

Format and typo checked the README.md too. This is the only entry in
1995 that I have done this for so far and is only done because I wanted
to make the alt code the suggested version to use first.