Fix 1992/westley/README.md about fixed bug
Removed the INABIAF section.

Added to alt section the subsections for build and to run. I did not add
an alternate try command because I would have to look up some
coordinates in the US (that's what it's for) and that takes time. I can
do that when I do a proper check of this file.