Disable -Wno-macro-redefined in 1986/wall/Makefile
This is because of what the entry is, the author's remarks and the fact
that the warning produces fun output for the entry itself. The author
had stated other warnings were fun but what these warnings were are no
longer known and the messages almost certainly would be different now.