Fix script in 1990/jaw
Now the files added to the archive will be extracted. It is unfortunate
but a temporary tarball has to be formed rather than writing to stdout.
There is a possible fix to this but it hasn't been looked into yet. The
tarball is temporary, however, and is deleted after it is extracted. The
script MUST be run from a subdirectory of the entry but as the script
had hardcoded the paths before this shouldn't be that big of a deal.

Another problem was more to do with paths and also the script relying on
the exit code allowing for &&. Now it's ';' between most commands.

There still remains, Landon and I believe, a bug in the entry and this
has been added to the bugs.md file; the script not working has been
removed from the bugs.md.