Update bugs.md for 1990/jaw fixes
There still remains another bug or at least it is thought it is a bug
and this has been added whereas the other one about scripts removed.
This was noted in commit 086588b1a0e0bf1aa87d5b0188bc72d49ddc08f6 but I
accidentally left it off in the commit.