Restore 2001/herrmann2 original code
By some chance it was in .gitignore and was entirely missing so I
restored it from the archive.

I also updated the README.md file about the alt (which is the original
in fact) not compiling with clang (something I fixed in the entry
itself) in addition to it might not working in some systems (like 64-bit
systems).