Minor update to thanks file
I've declared myself the IOCCC's resident corrections officer and
according to Landon I'm a fine one at that.

I also elaborated on the reason for fgets() change but also added a note
that in some cases it's not possible or ideal and in some cases it's not
easy especially due to the annoying fact that for 'compatibility'
reasons fgets() retains the newline and gets() does not.